environments/agent/requirements.txt
fastapi
uvicorn[standard]
openai
langchain
langchain-openai
langchain-community
python-dotenv